Move query string stuff to a separate module
